requirements.txt
Django==1.9.5
SPARQLWrapper==1.7.6
django-tastypie==0.13.3
django-widget-tweaks==1.4.1
html5lib==0.9999999
isodate==0.5.4
keepalive==0.5
pyparsing==2.1.1
python-dateutil==2.5.2
python-mimeparse==1.5.1
rdflib==4.2.1
six==1.10.0
